quell warning on mingw
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6740 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/CaptureReplay.cxx
@@ -366,7 +366,7 @@ routePacket (u16 code, int len, const void * data, bool fake)
 bool 
 Capture::addPacket (u16 code, int len, const void * data, bool fake)
 {
-  if ((getCRtime() - UpdateTime) > UpdateRate) {
+  if ((getCRtime() - UpdateTime) > (int)UpdateRate) {
     saveStates ();
   }
   return routePacket (code, len, data, fake);
